backend_safe/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.8.2
sqlalchemy==2.0.36
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
supabase==2.10.0

# Nota: O backend usa os mesmos módulos do projeto principal:
# - sheets_config.py
# - sheets_database.py  
# - models.py
# - database.py
# 
# Certifique-se de que credentials.json está na raiz do projeto
# ou configure as variáveis de ambiente GOOGLE_CREDENTIALS e GOOGLE_SHEET_ID
